OK-891: Refaktoroitu valintalaskennan tulosten näyttämisen yhteiset osat yhteiseen ValintalaskentaResult-komponenttiin
