+++ src/app/haku/[oid]/valintaryhma/[valintaryhma]/components/valintaryhma-valintalaskenta.tsx
@@ -2,16 +2,10 @@
 
 import { useTranslations } from '@/lib/localization/useTranslations';
 import { Divider, Stack } from '@mui/material';
-import {
-  LaskentaActorRef,
-  LaskentaState,
-  useLaskentaApi,
-} from '@/lib/state/laskenta-state';
+import { LaskentaActorRef, useLaskentaApi } from '@/lib/state/laskenta-state';
 import { Hakukohde } from '@/lib/kouta/kouta-types';
-import { ConfirmationModal } from '@/components/modals/confirmation-modal';
-import { SuorittamattomatHakukohteet } from '@/components/suorittamattomat-hakukohteet';
-import { LaskentaStateButton } from '@/components/laskenta-state-button';
-import { ValintalaskentaStatus } from '@/components/ValintalaskentaStatus';
+import { OphButton } from '@opetushallitus/oph-design-system';
+import { ValintalaskentaResult } from '@/components/ValintalaskentaResult';
 
 export const ValintaryhmanValintalaskenta = ({
   hakukohteet,
@@ -20,29 +14,22 @@ export const ValintaryhmanValintalaskenta = ({
   hakukohteet: Array<Hakukohde>;
   actorRef: LaskentaActorRef;
 }) => {
-  const { state, confirmLaskenta, cancelLaskenta } = useLaskentaApi(actorRef);
+  const { state, startLaskenta } = useLaskentaApi(actorRef);
 
   const { t } = useTranslations();
 
   return (
     <Stack spacing={2}>
-      <ConfirmationModal
-        title={t('valinnanhallinta.varmista')}
-        open={state.matches(LaskentaState.WAITING_CONFIRMATION)}
-        onConfirm={confirmLaskenta}
-        onCancel={cancelLaskenta}
-      />
-      <ValintalaskentaStatus
-        laskentaActorRef={actorRef}
+      {state.hasTag('stopped') && !state.hasTag('completed') && (
+        <OphButton variant="contained" onClick={startLaskenta}>
+          {t('valintalaskenta.suorita-valintalaskenta')}
+        </OphButton>
+      )}
+      <ValintalaskentaResult
+        actorRef={actorRef}
+        hakukohteet={hakukohteet}
         progressType="spinner"
       />
-      <LaskentaStateButton actorRef={actorRef} />
-      {state.hasTag('completed') && (
-        <SuorittamattomatHakukohteet
-          actorRef={actorRef}
-          hakukohteet={hakukohteet}
-        />
-      )}
       {(state.hasTag('started') || state.hasTag('completed')) && (
         <Divider sx={{ paddingTop: 1 }} />
       )}
